// my-app/src/Helpers/Enums.tsx
export enum Disciplines{
    Null = "-",
    BusinessModelling = "Modelowanie biznesowe",
    Requirements = "Gromadzenie wymagan",
    AnalysisAndDesign = "Analiza i projektowanie",
    Implementation = "Implementacja",
    Test = "Test",
    Deployment = "Dostarczanie/wdrazanie",
    ConfigurationAndChange = "Konfiguracja i zmiany",
    ProjectManagement = "Zarzadzanie projektami",
    Environment = "srodowisko"
}

export enum Categories{
    Micro = "Micro",
    Small = "Small",
    Medium = "Medium",
    Big = "Big",
    VeryBig = "VeryBig",
    Huge = "Huge"
}